requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
faiss-cpu
pydantic
python-dotenv
python-multipart
numpy
